Add Catalog Builder
Add a serverless service that builds a product catalog.
This service defines a dynamodb table for materializing our last best understanding of the product catalog into.
This service defines a lambda that processes product/create events on the stream to populate the catalog.

README.md describes the service
serverless.yml defines the service
package.json brings in the schema validator
catalog.js defines the code for the lambda
catalog.spec.js defines a shell into which we can pour test implementations

Temporary:
  copy of the retail-stream-schema-ingress.json schema
  copy of the product-create-schema.json