Lighten Lambda Package
Eliminate the dev dependencies for mocha and chai.  These are declared by the parent hello-retail project and are not needed here.  This leaves just the schemas and catalog.spec.js files as the only strictly unnecessary content in the Lambda package.  We may want to remove the schemas if we can source them dynamically but we may also not want to do that.  The catalog.spec.js file is light enough that I don't yet care.