Photographer Assignment, First Hack (#70)
* rename `execute` to `processor` to reflect more general charter

* Photographer Assignment, First Hack

Broadly speaking: Create a table that is updated idempotently with registrations of photographers.
Assignment is successful when a contingent write to that table succeeds.
Failure to fulfill the assignment removes the assignment.
Successful completion of the assignment removes the assignment, incrementing the assignments.

* Bug Fixes, Round 1

Assign: add catch block to report and handle unexpected errors
Unmessage: use the right name for the function... you know, so it gets called.  Also, improve error reporting just in case.

* Bug Fixes, Round 2

Assign: audit for dropped executions
Unmessage: fix update expression (two extra commas)

* remove personal info

* Use the correct table and table name

* Add missing eslint disable statement (how'd it pass pre-commit hooks)
Increase default capacity for the registered-photographers-by-completed-assignments GSI

* add forgotten npm install for 1.assign

* Bug Fixes, Round 7

Remove unused assignment table name from message lambda
Improve message when a text is received but an assignment doesn't exist
Make the handling of an unassignment via response and expiry consistent.  Delete the "pending assignment" regardless of it's invalidity
In so ^ doing, increase cleanliness of deletion in success case

* Bug Fixes, Round 8

Oops.  Forgot to forward the original event back into the StepFunction! :P
Also, update the permissions to match the last changes in code.

* Feature: Merchant Name

For the purpose of better identifying whose product it is, add the merchant's name to the text sent to the photographer.  This will enable them to better identify the participant who has the physical product

* Bug Fixes, Round 9

event is not the error, it's the success.  add null

* Bump Assignment Failure Wait Time

When an assignment is not immediately made, we wait.  This changes the amount of waiting from 5 seconds to 15.  A negligible amount from a human standpoint but this extends the amount of time the system can wait for a photographer to assign by a factor of three.  It cuts costs slightly but at the expected scale, this is negligible.